Update modal
updated the modal and how the info from that is parsed. Changed the payload to be saved per user ID. This should allow you to access your payload from any cog/button so it can be edited or reused. 

This will now not allow other users to use the info from your payload since its stored by the user ID.